collector/requirements.txt
# BRRR Capital Data Collector - Python Dependencies
requests>=2.28.0
python-dotenv>=1.0.0
signalrcore>=0.9.5
